+++ CHANGELOG.md
@@ -25,6 +25,7 @@ You can find and compare releases at the [GitHub release page](https://github.co
 - Exclude unused standard types from the schema
 - Require lazy type loader to return `Type` directly without an intermediary callable
 - Rename `ServerConfig` option `persistentQueryLoader` to `persistedQueryLoader`
+- Call previously unused methods `EnumType::parseValue()` and `EnumType::parseLiteral()`
 
 ### Added